api/requirements.txt
Flask
peewee
torch
Pillow
psycopg2-binary
gunicorn